started separating the ecore-specifics from the error extractor. This results in an object instead of a class to make it conform to an interface.
